fix: Environment provider should notify watchers once all items have been added
Currently, when an environment provider is registered, the notify event
is sent when the provider is added to a store. If a consumer is
consuming the event right away, the items may not have been added to the
inmemory store yet, then leading to a bad behaviour.

We now re-notify watchers that configstore.Store items have been changed
when all items are added.

Signed-off-by: Romain Beuque <556072+rbeuque74@users.noreply.github.com>